Add windowframes and windows to manage map display
So far we are only using the primary window (level_map) as the "playing
area" and the subwindow (viewport) as the visible area. Currently the
game generates a map equal in size to the user's terminal, and clips
off a bit at the bottom and the sides so we can pretend there is a HUD
there. Coming soon (tm)!